requirements.txt
Flask==2.3.2
graphene==2.1.9
Flask-GraphQL==2.0.1
graphql-core==2.3.2
image==1.5.33
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
google-cloud-vision==3.4.0
Pillow==10.0.0
google-cloud==0.34.0
google-cloud-vision==3.4.0

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

gunicorn